moved constant into shared
